return 403 for pos old apis if partner already migrated
Type(s): UPDATE

Issue(s):
- JIRA:
https://blabol.atlassian.net/browse/MSBE-2078?atlOrigin=eyJpIjoiMzdmYjU5YzM5ZGUyNGJiYzgxNTYyZDYyMTFmYjkwYmYiLCJwIjoiaiJ9

+++ app/Http/Controllers/Pos/OrderController.php
@@ -159,6 +159,8 @@ class OrderController extends Controller
      */
     public function store($partner, Request $request, Creator $creator, ProfileCreator $profileCreator, PosCustomerCreator $posCustomerCreator, PartnerRepository $partnerRepository, PaymentLinkCreator $paymentLinkCreator)
     {
+        $partner = $request->partner;
+        if($partner->isMigrated(Modules::POS))  return api_response($request, null, 403);
         $this->validate($request, [
             'services' => 'required|string',
             'paid_amount' => 'sometimes|required|numeric',
@@ -274,6 +276,8 @@ class OrderController extends Controller
     public function delete($partner, $order, Request $request, PosOrderDeleter $deleter)
     {
         try {
+            $partner = $request->partner;
+            if($partner->isMigrated(Modules::POS))  return api_response($request, null, 403);
             $deleter->setPartner($request->partner)->setOrder($order)->delete();
             return api_response($request, true, 200);
         } catch (PosExpenseCanNotBeDeleted $e) {
@@ -323,6 +327,8 @@ class OrderController extends Controller
      */
     public function update(Request $request, Updater $updater, InvoiceService $invoiceService)
     {
+        $partner = $request->partner;
+        if($partner->isMigrated(Modules::POS))  return api_response($request, null, 403);
         $this->setModifier($request->manager_resource);
             /** @var PosOrder $order */
             $new           = 1;
@@ -350,6 +356,8 @@ class OrderController extends Controller
      */
     public function updateStatus(Request $request, StatusChanger $statusChanger)
     {
+        $partner = $request->partner;
+        if($partner->isMigrated(Modules::POS))  return api_response($request, null, 403);
         $this->setModifier($request->manager_resource);
         $order = PosOrder::with('items')->find($request->order);
         $statusChanger->setOrder($order)->setStatus($request->status)->setModifier($request->manager_resource)->changeStatus();
